cms/pyproject.toml
[project]
name = "cms"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "matv864", email = "matv864@gmail.com" }
]
dependencies = [
    "sqladmin>=0.20.1",
    "asyncpg>=0.30.0",
    "fastapi>=0.115.11",
    "granian[reload]>=2.0.1",
    "python-dotenv>=1.0.1",
    "itsdangerous>=2.2.0",
]
readme = "README.md"
requires-python = ">= 3.12"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/cms"]
